DocInsights/requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.2
openpyxl>=3.1.2
PyPDF2>=3.0.0
pdfplumber>=0.10.2
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0